% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_binary_params}
\alias{format_binary_params}
\title{Function that selects the toolbar options that are either TRUE/FALSE to JSON format}
\usage{
format_binary_params(params)
}
\arguments{
\item{params}{Toolbar options selected by the user, the output of `toolbar_options()`}
}
\value{
a json chr with the toolbar options that were set to TRUE
}
\description{
Function that selects the toolbar options that are either TRUE/FALSE to JSON format
}
