% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_multiple_choice_options}
\alias{format_multiple_choice_options}
\title{Function that transforms a toolbar option that is a vector/list of choices to JSON format}
\usage{
format_multiple_choice_options(name, value)
}
\arguments{
\item{name}{The name of the choice, e.g. "align"}

\item{value}{The value(s) of the choice, e.g. c("center", "right")}
}
\value{
a json chr with the toolbar option and its values
}
\description{
Function that transforms a toolbar option that is a vector/list of choices to JSON format
}
