% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_toolbar_options}
\alias{format_toolbar_options}
\title{It formats the toolbar options selected by the user into a JSON format}
\usage{
format_toolbar_options(params)
}
\arguments{
\item{params}{Toolbar options selected by the user, the output of `toolbar_options()`.}
}
\value{
A list containing all the choices for the toolbar in JSON format.
}
\description{
It formats the toolbar options selected by the user into a JSON format
}
