% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quill.R
\name{quill}
\alias{quill}
\title{Create a text editor}
\usage{
quill(
  quill_params = quill_options(),
  width = NULL,
  height = NULL,
  theme = "snow",
  placeholder = "Placeholder message",
  elementId = NULL
)
}
\arguments{
\item{quill_params}{Use the function `quill_options()` to define the params for the toolbar.}

\item{width}{The width of the editor}

\item{height}{The height of the editor}

\item{theme}{Either 'snow' or 'bubble' are the two themes p}

\item{placeholder}{Placeholder text that will appear in the editor}

\item{elementId}{Use an explicit element ID for the widget
(rather than an automatically generated one).}
}
\description{
This function creates a QuillJS text editor widget using \pkg{htmlwidgets}. The
widget can be rendered on HTML pages generated from R Markdown, Shiny, or
other applications.
}
\examples{
library(shiny)
library(quill)
ui = shinyUI(fluidPage(
  div(
    div(
      "This is your new text editor",
      quillOutput("editor_1")
    ),
    div(
      "You can easily render what the user is writing as HTML",
      htmlOutput('rendered_output')
    )
  )
))

server = function(input, output) {
  output$editor_1 <- renderQuill({
    quill(
      quill_params = quill_options()
    )
  })
  observeEvent(input$editor_1, {
    cat(
      sep = "\n",
      get_editor_content(input$editor_1, format = "JSON"),
      get_editor_content(input$editor_1, format = "HTML"),
      get_editor_content(input$editor_1, format = "Text")
    )
  })
  output$rendered_output <- shiny::renderUI(HTML(get_editor_content(input$editor_1, format = "HTML")))
}

shinyApp(ui = ui, server = server)
}
