% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quill.R
\name{quill}
\alias{quill}
\title{Create a text editor}
\usage{
quill(
  toolbar = toolbar_options(),
  width = NULL,
  height = NULL,
  theme = "snow",
  placeholder = "Placeholder message",
  elementId = NULL
)
}
\arguments{
\item{toolbar}{Use the function `toolbar_options()` to define the parameters of the toolbar.}

\item{width}{The width of the editor.}

\item{height}{The height of the editor.}

\item{theme}{Either 'snow' or 'bubble' are the supported themes.}

\item{placeholder}{Placeholder text that will appear in the editor.}

\item{elementId}{Use an explicit element ID for the widget
(rather than an automatically generated one).}
}
\value{
A HTML widget object.
}
\description{
This function creates a QuillJS text editor widget using \pkg{htmlwidgets}. The
widget can be rendered on HTML pages generated from R Markdown, Shiny, or
other applications.
}
\examples{
quill(toolbar = toolbar_options(align = 'right'))

}
