% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toolbar_options}
\alias{toolbar_options}
\title{Configure the Toolbar Options}
\usage{
toolbar_options(
  align = c("center", "right"),
  background = c(),
  blockquote = TRUE,
  bold = TRUE,
  clean = TRUE,
  color = c(),
  direction = "rtl",
  formula = TRUE,
  code = TRUE,
  italic = TRUE,
  image = TRUE,
  header = list(c(1, 2, "normal_selected", 3)),
  indent = c("-1", "+1"),
  link = TRUE,
  list = c("ordered", "bullet", "check"),
  size = c("small", "normal_unselected", "large", "huge"),
  strike = TRUE,
  script = c("sub", "super"),
  underline = TRUE,
  video = FALSE,
  code_block = TRUE
)
}
\arguments{
\item{align}{A vector containing types of alignment, e.g., `c('center', 'right')`.}

\item{background}{A vector of colors for the background. If empty, default colors will appear,
e.g., `c('red', 'blue', '#32a852')` or `c()`.}

\item{blockquote}{Logical; `TRUE` or `FALSE` to include/exclude blockquote functionality.}

\item{bold}{Logical; `TRUE` or `FALSE` to include/exclude bold functionality.}

\item{clean}{Logical; `TRUE` or `FALSE` to include/exclude clean functionality.}

\item{color}{A vector of colors for text color. If empty, default colors will appear,
e.g., `c('red', 'blue', '#32a852')` or `c()`.}

\item{direction}{Character; `NULL` or `"rtl"` to set the text input direction.}

\item{formula}{Logical; `TRUE` or `FALSE` to include/exclude formula functionality.}

\item{code}{Logical; `TRUE` or `FALSE` to include/exclude inline code functionality.}

\item{italic}{Logical; `TRUE` or `FALSE` to include/exclude italic functionality.}

\item{image}{Logical; `TRUE` or `FALSE` to include/exclude image embedding functionality.}

\item{header}{A vector or list for header options. Example: `c(1, 2)` creates H1 and H2 buttons.
Use `list(c(1, 2, 3, "normal_selected"/"normal_unselected"))` for a collapsed dropdown with H1,
H2, and H3 and 'Normal'.
`normal_selected" / "normal_unselected"` controls the "Normal" button's inclusion, with
 selection depending on the current default.
`TRUE`/`FALSE` can add or exclude the "Normal" button; `header = c(1, 2, 3, TRUE)`.}

\item{indent}{A vector for indentation controls, e.g., `c('-1', '+1')` or `NULL`.}

\item{link}{Logical; `TRUE` or `FALSE` to include/exclude hyperlink functionality.}

\item{list}{A vector specifying list types, e.g., `c("ordered", "bullet", "check")`.}

\item{size}{A vector specifying size options: `c('small', "normal_selected"/"normal_unselected",
'large', 'huge')`.
 `normal_selected" / "normal_unselected"` controls the "Normal" button's inclusion, with
 selection depending on the current default.}

\item{strike}{Logical; `TRUE` or `FALSE` to include/exclude strikethrough functionality.}

\item{script}{A vector for script options, e.g., `c("sub", "super")`.}

\item{underline}{Logical; `TRUE` or `FALSE` to include/exclude underline functionality.}

\item{video}{Logical; `TRUE` or `FALSE` to include/exclude video functionality.}

\item{code_block}{Logical; `TRUE` or `FALSE` to include/exclude code block functionality.}
}
\value{
list
}
\description{
This function configures toolbar options for the editor, allowing customization of
formatting features such as alignment, color, font styles, and more.
}
\examples{
quill(toolbar = toolbar_options(align = 'right', bold = FALSE))
}
